agent_pgv/requirements.txt
google-generativeai
python-dotenv
langchain
PyPDF2
langchain_google_genai
langchain-community
python-multipart
psycopg2-binary 
uvicorn
sqlalchemy
pgvector
fastapi
pandas